% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_last_consecutive_signs.R
\name{count_last_consecutive_signs}
\alias{count_last_consecutive_signs}
\title{Number of last consecutive returns with same sign}
\usage{
count_last_consecutive_signs(ticker, start_date, end_date, frequency)
}
\arguments{
\item{ticker}{Stock symbol on Yahoo Finance.}

\item{start_date}{Start date of the data.}

\item{end_date}{End date of the data.}

\item{frequency}{Periodicity of the data (e.g., "daily", "weekly").}
}
\value{
The number of times this event occurs.
}
\description{
This function computes the number of last consecutive strictly positive or
negative returns for a given stock.
}
\examples{
count_last_consecutive_signs("AAPL", "2022-01-01", "2022-12-31", "daily")

}
