% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_consecutive_returns.R
\name{number_consecutive_returns}
\alias{number_consecutive_returns}
\title{Number of consecutive returns with same sign}
\usage{
number_consecutive_returns(
  n,
  ticker,
  start_date,
  end_date,
  frequency,
  positive = TRUE
)
}
\arguments{
\item{n}{Number of consecutive returns with same sign.}

\item{ticker}{Stock symbol on Yahoo Finance.}

\item{start_date}{Start date of the data.}

\item{end_date}{End date of the data.}

\item{frequency}{Periodicity of the data (e.g., "daily", "weekly").}

\item{positive}{Boolean which gives consecutive positive returns if TRUE (by default) or
negative if FALSE.}
}
\value{
The number of times this event occurs.
}
\description{
This function computes the number consecutive strictly positive or negative
returns for a given stock.
}
\examples{
number_consecutive_returns(3, "AAPL", "2022-01-01", "2022-12-31", "daily", positive = FALSE)

}
