% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prices_consecutive_returns.R
\name{prices_consecutive_returns}
\alias{prices_consecutive_returns}
\title{Stock price in correspondence of consecutive returns with same sign}
\usage{
prices_consecutive_returns(
  n,
  ticker,
  start_date,
  end_date,
  frequency,
  positive = TRUE
)
}
\arguments{
\item{n}{Number of consecutive returns with same sign.}

\item{ticker}{Stock symbol on Yahoo Finance.}

\item{start_date}{Start date of the data.}

\item{end_date}{End date of the data.}

\item{frequency}{Periodicity of the data (e.g., "daily", "weekly").}

\item{positive}{Boolean which gives consecutive positive returns if TRUE (by default) or
negative if FALSE.}
}
\value{
The stock price when this event occurs.
}
\description{
This function computes the stock price when consecutive strictly positive or
negative returns occur.
}
\examples{
prices_consecutive_returns(3, "AAPL", "2022-01-01", "2022-12-31", "daily", positive = TRUE)

}
