% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_consecutive_returns.R
\name{probability_consecutive_returns}
\alias{probability_consecutive_returns}
\title{Probability of consecutive returns with the same sign according to historical data}
\usage{
probability_consecutive_returns(
  n,
  ticker,
  start_date,
  end_date,
  frequency,
  positive = TRUE
)
}
\arguments{
\item{n}{Number of consecutive returns with same sign.}

\item{ticker}{Stock symbol on Yahoo Finance.}

\item{start_date}{Start date of the data.}

\item{end_date}{End date of the data.}

\item{frequency}{Periodicity of the data (e.g., "daily", "weekly").}

\item{positive}{Boolean which gives consecutive positive returns if TRUE (by default) or
negative if FALSE.}
}
\value{
The percentage of times this event occurs.
}
\description{
This function computes the probability of consecutive strictly positive or
negative returns for a given stock.
}
\examples{
probability_consecutive_returns(3, "AAPL", "2022-01-01", "2022-12-31", "daily",
positive = FALSE)

}
